Section 19, Lesson 269 - Converting CategoryVC to subclass SwipeTableVC
Converted the CategoryVC to be a subclass of the SwipeTableVC.
Changed the identifier from CategoryCell (specific) to SwipeableCell
(generic). Put most of the SwipeCellKit functionality inside the
SwipeTableVC class (super), and overrided methods where necessary
in the CategoryVC.
