+++ xcode-projects/xcode-todoey-app/Todoey/Todoey/Controllers/SwipeTableViewController.swift
@@ -7,30 +7,49 @@
 //
 
 import UIKit
+import SwipeCellKit
 
-class SwipeTableViewController: UITableViewController {
-
+class SwipeTableViewController: UITableViewController, SwipeTableViewCellDelegate {
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-
     }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-    // MARK: - Table view data source
-
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        // #warning Incomplete implementation, return the number of sections
-        return 0
-    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
-        return 0
-    }
-
     
+    // MARK: - SwipeCell Delegate Methods
+    
+    func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath, for orientation: SwipeActionsOrientation) -> [SwipeAction]? {
+        guard orientation == .right else { return nil }
+        
+        let deleteAction = SwipeAction(style: .destructive, title: "Delete") { action, indexPath in
+            // handle action by updating model with deletion
+            print("Delete Cell")
+            self.updateModel(at: indexPath)
+
+        }
+        
+        // customize the action appearance
+        deleteAction.image = UIImage(named: "delete")
+        
+        return [deleteAction]
+    }
+    
+    func tableView(_ tableView: UITableView, editActionsOptionsForRowAt indexPath: IndexPath, for orientation: SwipeActionsOrientation) -> SwipeTableOptions {
+        var options = SwipeTableOptions()
+        options.expansionStyle = .destructive
+        options.transitionStyle = .border
+        return options
+    }
+    
+    // MARK: TableView Datasource Methods
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "SwipeableCell", for: indexPath) as! SwipeTableViewCell
+        
+        cell.delegate = self
+        
+        return cell
+    }
+    
+    func updateModel(at indexPath: IndexPath) {
+        // Update DataModel
+    }
 }
